Reorganisation of code
